Rename the authnext module to authnonext which doesn't sound like the word "next".
